requirements_tests.txt
# requirements_tests.txt
# Add testing dependencies here

pytest
pytest-cov
pytest-homeassistant-custom-component
